Cybersecurity_Agent/requirements.txt
langchain
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
gradio
langchain
langchain-google-genai
gradio
python-dotenv